Update README
Fix some typos, update some things, and prepare for a more cross-platform and agnostic approach to how we write our documentation.

Everything compiles and works on Mac and Linux, and on any modern browser (I have tried many). Windows can use either the Linux Subsystem For Windows or Cygwin/MSYS/MinGW for an equivalent installation process. If you want to do it on Windows natively, I'm not sure exactly how to accomplish that without using Cygwin/MSYS/MinGW/Linux Subsystem For Windows.

Java is still an incredibly unsafe and buggy piece of software, and I don't recommend anybody install it on any operating system. The web plugin is no longer supported in many web browsers because of this, so unless you are developing Java software, don't install it.

@schibo

Tell me if there is anything you want to change or update or anything. :smile: